# models/index.js
const Department = require("@/models/department");
const Program = require("@/models/program");
const User = require("@/models/user");
const Course = require("@/models/course");
const AcademicRecord = require("@/models/academicRecord");
const Attendance = require("@/models/attendance");

module.exports = {
  Department,
  Program,
  User,
  Course,
  AcademicRecord,
  Attendance,
};