# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "School Management System API",
    "version": "1.0.0",
    "description": "API documentation for the School Management System"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
  "paths": {
    "/academic/{id}": {
      "get": {
        "summary": "Get an academic record by ID",
        "description": "Retrieve a single academic record by its ID",
        "tags": [
          "Academic Records"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with academic record details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AcademicRecord"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "summary": "Update an academic record",
        "description": "Update an academic record's information by its ID",
        "tags": [
          "Academic Records"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRecordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Academic record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Academic record updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AcademicRecord"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "summary": "Delete an academic record",
        "description": "Delete an academic record by its ID",
        "tags": [
          "Academic Records"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Academic record deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Academic record deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/academic": {
      "get": {
        "summary": "Get all academic records",
        "description": "Retrieve a list of all academic records with optional filtering, sorting, pagination, and search capabilities",
        "tags": [
          "Academic Records"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/sortParam"
          },
          {
            "$ref": "#/components/parameters/fieldsParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/limitParam"
          },
          {
            "$ref": "#/components/parameters/searchParam"
          },
          {
            "in": "query",
            "name": "studentId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by student ID"
          },
          {
            "in": "query",
            "name": "courseId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by course ID"
          },
          {
            "in": "query",
            "name": "grade[gte]",
            "schema": {
              "type": "number"
            },
            "description": "Filter by grade greater than or equal to value"
          },
          {
            "in": "query",
            "name": "grade[lte]",
            "schema": {
              "type": "number"
            },
            "description": "Filter by grade less than or equal to value"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with list of academic records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "example": 10
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/AcademicRecord"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "summary": "Create a new academic record",
        "description": "Create a new academic record with the provided data",
        "tags": [
          "Academic Records"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcademicRecordInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Academic record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Academic record created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AcademicRecord"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Grade": {
        "type": "object",
        "properties": {
          "midterm": {
            "type": "number",
            "example": 85,
            "description": "Midterm exam grade"
          },
          "final": {
            "type": "number",
            "example": 92,
            "description": "Final exam grade"
          },
          "assignments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Homework 1",
                  "description": "Name of the assignment"
                },
                "score": {
                  "type": "number",
                  "example": 90,
                  "description": "Score received for the assignment"
                },
                "weight": {
                  "type": "number",
                  "example": 20,
                  "description": "Weight of the assignment in percentage"
                }
              }
            }
          },
          "totalScore": {
            "type": "number",
            "example": 90,
            "description": "Calculated total score"
          },
          "letterGrade": {
            "type": "string",
            "enum": [
              "A",
              "A-",
              "B+",
              "B",
              "B-",
              "C+",
              "C",
              "C-",
              "D+",
              "D",
              "F"
            ],
            "example": "A",
            "description": "Letter grade equivalent"
          }
        }
      },
      "AcademicRecord": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "60d21b4667d0d8992e610c92"
          },
          "studentId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c89"
              },
              "username": {
                "type": "string",
                "example": "john.doe"
              },
              "firstName": {
                "type": "string",
                "example": "John"
              },
              "lastName": {
                "type": "string",
                "example": "Doe"
              },
              "UserId": {
                "type": "string",
                "example": "JD123"
              }
            },
            "description": "Reference to the student"
          },
          "courseId": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "example": "60d21b4667d0d8992e610c91"
              },
              "courseCode": {
                "type": "string",
                "example": "CS101"
              },
              "title": {
                "type": "string",
                "example": "Introduction to Computer Science"
              }
            },
            "description": "Reference to the course"
          },
          "semester": {
            "type": "string",
            "example": "Fall",
            "description": "Academic semester (e.g., Fall, Spring, Summer)"
          },
          "academicYear": {
            "type": "string",
            "example": "2024-2025",
            "description": "Academic year (e.g., \"2024-2025\")"
          },
          "registrationStatus": {
            "type": "string",
            "enum": [
              "registered",
              "dropped",
              "completed",
              "failed",
              "withdrawn"
            ],
            "example": "registered",
            "description": "Current registration status for the course"
          },
          "grade": {
            "type": "object",
            "$ref": "#/components/schemas/Grade",
            "description": "Grade information for the course"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "studentId",
          "courseId",
          "semester",
          "academicYear",
          "registrationStatus"
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "example": 100,
            "description": "Total number of records"
          },
          "page": {
            "type": "integer",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "integer",
            "example": 25,
            "description": "Number of records per page"
          },
          "pages": {
            "type": "integer",
            "example": 4,
            "description": "Total number of pages"
          },
          "hasNext": {
            "type": "boolean",
            "example": true,
            "description": "Whether there is a next page"
          },
          "hasPrev": {
            "type": "boolean",
            "example": false,
            "description": "Whether there is a previous page"
          }
        }
      },
...
    },
    "parameters": {
      "sortParam": {
        "in": "query",
        "name": "sort",
        "schema": {
          "type": "string"
        },
        "description": "Sort order (prefix field with - for descending) e.g. -createdAt,name"
      },
      "fieldsParam": {
        "in": "query",
        "name": "fields",
        "schema": {
          "type": "string"
        },
        "description": "Fields to include (comma separated) e.g. name,code,description"
      },
      "pageParam": {
        "in": "query",
        "name": "page",
        "schema": {
          "type": "integer",
          "default": 1
        },
        "description": "Page number for pagination"
      },
      "limitParam": {
        "in": "query",
        "name": "limit",
        "schema": {
          "type": "integer",
          "default": 25,
          "maximum": 100
        },
        "description": "Number of records per page (max 100)"
      },
      "searchParam": {
        "in": "query",
        "name": "search",
        "schema": {
          "type": "string"
        },
        "description": "Search term for full-text search"
      },
      "idParam": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Resource ID"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - validation failed",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Validation failed"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": [
                    "Name is required",
                    "Invalid status value"
                  ]
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}